ddpm works on celeba

+++ models/utils.py
@@ -118,10 +118,10 @@ def get_model_fn(model, train=False):
     """
     if not train:
       model.eval()
-      return model(x, labels)
+      return model(x, labels).type(torch.float32)
     else:
       model.train()
-      return model(x, labels)
+      return model(x, labels).type(torch.float32)
 
   return model_fn
